requirements.txt
streamlit~=1.22.0
numpy~=1.23.5
Pillow~=9.5.0
tensorflow-cpu
pandas~=2.0.1